% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ace-constant-columns.R
\docType{data}
\name{ace_header}
\alias{ace_header}
\alias{COL_FILE}
\alias{Q_COL_FILE}
\alias{COL_TIME}
\alias{Q_COL_TIME}
\alias{COL_N_FINISHED}
\alias{Q_COL_N_FINISHED}
\alias{COL_RT}
\alias{Q_COL_RT}
\alias{COL_CORRECT_BUTTON}
\alias{Q_COL_CORRECT_BUTTON}
\alias{COL_PREV_CORRECT_BUTTON}
\alias{Q_COL_PREV_CORRECT_BUTTON}
\alias{COL_CORRECT_RESPONSE}
\alias{Q_COL_CORRECT_RESPONSE}
\alias{COL_RW}
\alias{Q_COL_RW}
\alias{COL_LATE_RESPONSE}
\alias{Q_COL_LATE_RESPONSE}
\alias{COL_PREV_LATE_RESPONSE}
\alias{Q_COL_PREV_LATE_RESPONSE}
\alias{COL_CONDITION}
\alias{Q_COL_CONDITION}
\alias{COL_STUDY_COND}
\alias{Q_COL_STUDY_COND}
\alias{COL_PID}
\alias{Q_COL_PID}
\alias{COL_BID}
\alias{Q_COL_BID}
\alias{COL_BID_SHORT}
\alias{Q_COL_BID_SHORT}
\alias{COL_NAME}
\alias{Q_COL_NAME}
\alias{COL_AGE}
\alias{Q_COL_AGE}
\alias{COL_GRADE}
\alias{Q_COL_GRADE}
\alias{COL_GENDER}
\alias{Q_COL_GENDER}
\alias{COL_HANDEDNESS}
\alias{Q_COL_HANDEDNESS}
\alias{COL_SUB_ID}
\alias{Q_COL_SUB_ID}
\alias{COL_PRACTICE}
\alias{Q_COL_PRACTICE}
\alias{COL_PRACTICE_RD}
\alias{Q_COL_PRACTICE_RD}
\alias{COL_PRACTICE_COUNT}
\alias{Q_COL_PRACTICE_COUNT}
\alias{COL_TRIAL_TYPE}
\alias{Q_COL_TRIAL_TYPE}
\alias{COL_TRIAL_NUM}
\alias{Q_COL_TRIAL_NUM}
\alias{COL_BLOCK_HALF}
\alias{Q_COL_BLOCK_HALF}
\alias{ALL_POSSIBLE_DEMOS}
\alias{Q_ALL_POSSIBLE_DEMOS}
\alias{ALL_POSSIBLE_EXPLORE_DEMOS}
\alias{Q_ALL_POSSIBLE_EXPLORE_DEMOS}
\alias{standardize_ace_column_names}
\alias{standardize_ace_column_types}
\alias{standardize_ace_values}
\title{ACE column name constants}
\format{
An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 1.

An object of class \code{name} of length 1.

An object of class \code{character} of length 8.

An object of class \code{list} of length 8.

An object of class \code{character} of length 23.

An object of class \code{list} of length 23.
}
\usage{
COL_FILE

Q_COL_FILE

COL_TIME

Q_COL_TIME

COL_N_FINISHED

Q_COL_N_FINISHED

COL_RT

Q_COL_RT

COL_CORRECT_BUTTON

Q_COL_CORRECT_BUTTON

COL_PREV_CORRECT_BUTTON

Q_COL_PREV_CORRECT_BUTTON

COL_CORRECT_RESPONSE

Q_COL_CORRECT_RESPONSE

COL_RW

Q_COL_RW

COL_LATE_RESPONSE

Q_COL_LATE_RESPONSE

COL_PREV_LATE_RESPONSE

Q_COL_PREV_LATE_RESPONSE

COL_CONDITION

Q_COL_CONDITION

COL_STUDY_COND

Q_COL_STUDY_COND

COL_PID

Q_COL_PID

COL_BID

Q_COL_BID

COL_BID_SHORT

Q_COL_BID_SHORT

COL_NAME

Q_COL_NAME

COL_AGE

Q_COL_AGE

COL_GRADE

Q_COL_GRADE

COL_GENDER

Q_COL_GENDER

COL_HANDEDNESS

Q_COL_HANDEDNESS

COL_SUB_ID

Q_COL_SUB_ID

COL_PRACTICE

Q_COL_PRACTICE

COL_PRACTICE_RD

Q_COL_PRACTICE_RD

COL_PRACTICE_COUNT

Q_COL_PRACTICE_COUNT

COL_TRIAL_TYPE

Q_COL_TRIAL_TYPE

COL_TRIAL_NUM

Q_COL_TRIAL_NUM

COL_BLOCK_HALF

Q_COL_BLOCK_HALF

ALL_POSSIBLE_DEMOS

Q_ALL_POSSIBLE_DEMOS

ALL_POSSIBLE_EXPLORE_DEMOS

Q_ALL_POSSIBLE_EXPLORE_DEMOS

standardize_ace_column_names(df, data_type)

standardize_ace_column_types(df)

standardize_ace_values(df, data_type)
}
\description{
ACE column name constants
}
\keyword{datasets}
\keyword{internal}
