% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-modules-ace.R, R/module-proc.R
\name{ace_procs}
\alias{ace_procs}
\alias{module_adp}
\alias{module_boxed}
\alias{module_brt}
\alias{module_colorselection}
\alias{module_flanker}
\alias{module_saat}
\alias{module_stroop}
\alias{module_spatialspan}
\alias{module_taskswitch}
\alias{module_tnt}
\alias{module_backwardsspatialspan}
\alias{module_filter}
\alias{module_ishihara}
\alias{module_spatialcueing}
\title{ACE/SEA module procs}
\usage{
module_adp(df)

module_boxed(df)

module_brt(df, data_type)

module_colorselection(df)

module_flanker(df)

module_saat(df, app_type)

module_stroop(df)

module_spatialspan(df)

module_taskswitch(df)

module_tnt(df)

module_backwardsspatialspan(df)

module_filter(df)

module_ishihara(df)

module_spatialcueing(df)
}
\description{
Methods for processing user data by \emph{group}.
}
\keyword{internal}
