% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-span.R
\name{ace_span}
\alias{ace_span}
\title{Returns the highest value that occurs at least 2x consecutively}
\usage{
ace_span(x)
}
\description{
Returns the highest value that occurs at least 2x consecutively
}
\keyword{internal}
