% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-turns.R
\name{ace_turns}
\alias{ace_turns}
\title{Average of the last 3 "turn around trials" - a correct trial that was preceded by an incorrect trial}
\usage{
ace_turns(x, y)
}
\description{
Average of the last 3 "turn around trials" - a correct trial that was preceded by an incorrect trial
}
\keyword{internal}

