\name{acer.estimate_acer}
\alias{acer.estimate_acer}
\title{Estimate the ACER function (internal)}
\description{A function the estimate ...}
\usage{acer.estimate_acer (x, eta, k)}
\arguments{
  \item{x}{a vector containing the data to be analyzed.}
  \item{eta}{the barrier levels.}
  \item{k}{the value of k for which the ACER function is to be
    estimated.}
}
\value{The estimated ACER function.}