\name{acer.evaluate}
\alias{acer.evaluate}
\title{Evaluate the ACER function (internal)}
\description{This function evaluated the ACER function at given levels.}
\usage{acer.evaluate (eta, par)}
\arguments{
  \item{eta}{barrier level(s) for which the ACER function is to be
    evaluated.}
  \item{par}{a list containing the parameters of the ACER function.}
}
\value{the values of the ACER function at the given barrier levels.}