\name{acer.linear_fit}
\alias{acer.linear_fit}
\title{Perform a linear fit on the estimated ACER function (internal)}
\description{...}
\usage{acer.linear_fit (eta, epsilon, peaks_mean)}
\arguments{
  \item{eta}{barrier_levels}
  \item{epsilon}{the values of the estimated ACER function corresponding
    to the barrier levels.}
  \item{peaks_mean}{the mean of the data used in the analysis.}
}
\value{
  \item{par}{A list containing the parameters of the ACER function from
    the linear fit.}
  }
