\name{acer.log_evaluate}
\alias{acer.log_evaluate}
\title{Evaluate the ACER function on the log scale (internal)}
\description{This function evaluated the ACER function on the log scale at given levels.}
\usage{acer.log_evaluate (eta, par)}
\arguments{
  \item{eta}{barrier level(s) for which the ACER function is to be
    evaluated.}
  \item{par}{a list containing the parameters of the ACER function.}
}
\value{the values of the ACER function at the given barrier levels.}