\name{acer.peaks}
\alias{acer.peaks}
\title{Extract peaks from a vector}
\description{This function extracts peaks (values which are preceded and
followed by values smaller than itself) from a vector, and returns those
peaks as well as the indices where the peaks were found.}
\usage{acer.peaks (x)}
\arguments{
  \item{x}{a numeric vector.}
}
\value{
  \item{peaks}{the peak values.}
}