\name{acer.plot_all_estimates}
\alias{acer.plot_all_estimates}
\title{Plot all estimates of the ACER functions (internal)}
\description{This function makes a plot of the estimated ACER function
  with various k levels}
\usage{acer.plot_all_estimates (eta, epsilon, k_values)}
\arguments{
  \item{eta}{barrier levels.}
  \item{epsilon}{a matrix containing estimates of the ACER function of
    a given k value in the corresponding column}
  \item{k_values}{the values of \eqn{k} to be plotted.}
  }