\name{acer.plot_estimate}
\alias{acer.plot_estimate}
\title{Plot an estimate of a single ACER function}
\description{This function plots the estimate of a single ACER function
  with confidence intervals.}
\usage{acer.plot_estimate (estimates, k_value)}
\arguments{
  \item{estimates}{an object of the class "acer.estimates" containing estimates of the ACER functions, returned from acer.estimate_functions.}
  \item{k_value}{the \eqn{k} value of the ACER function to be plotted.}
  }