\name{acer.plot_result}
\alias{acer.plot_result}
\title{Plot the results of the ACER analysis (internal)}
\description{This function plots the fitted ACER function with the
  corresponding confidence bounds.}
\usage{acer.plot_result (result)}
\arguments{
  \item{result}{an object of the class "acer.result".}
}
