\name{acer.print_to_screen}
\alias{acer.print_to_screen}
\title{Print results to screen (internal)}
\description{Prints the results of the ACER analysis to screen}
\usage{acer.print_to_screen (result)}
\arguments{
  \item{result}{a list of the class "acer.result" containing the results
    of an ACER analysis.}
}