% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aceR-package.R
\name{aceR_sample_data_path}
\alias{aceR_sample_data_path}
\title{aceR sample data path}
\usage{
aceR_sample_data_path(subfolder)
}
\arguments{
\item{subfolder}{which subfolder (and task type) do you want to test?
One of \code{c("email", "pulvinar","explorer", "sea")}. Must be specified.}
}
\value{
Returns the path containing sample aceR data for the
desired subfolder
}
\description{
aceR sample data path
}
