% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-write.R
\name{export_csv}
\alias{export_csv}
\title{Export CSV}
\usage{
export_csv(dat, path = ".", showWarnings = FALSE)
}
\arguments{
\item{dat}{a list of data frames to save}

\item{path}{the named release directory}

\item{showWarnings}{logical; should the warnings on failure be shown?}
}
\description{
Exports a list of data frames as csv into named directory
}

