% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-write.R
\name{export_plot}
\alias{export_plot}
\title{Export plot}
\usage{
export_plot(plot, file)
}
\arguments{
\item{plot}{a plot}

\item{file}{a character string giving the name of the file.}
}
\description{
Write plot to file
}

