% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-generic.R
\name{files_in_directory}
\alias{files_in_directory}
\title{List the files in a directory/folder}
\usage{
files_in_directory(path = ".", pattern = ".csv", recursive = TRUE)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link{getwd}()}.  Tilde
    expansion (see \code{\link{path.expand}}) is performed.  Missing
    values will be ignored.}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{recursive}{logical.  Should the listing recurse into directories?}
}
\value{
Returns a character vector containing the names of the files 
  in the specified directory.
}
\description{
Wrapper around \code{\link{list.files}()}
}
