% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-plots.R
\name{generate_plots_and_stats}
\alias{generate_plots_and_stats}
\title{Generate the Rmd ready plots and stats summaries}
\usage{
generate_plots_and_stats(df, x, stat_type, cohort = NULL)
}
\arguments{
\item{df}{The dataframe}

\item{x}{X-axis boxplot variable}

\item{stat_type}{the statistic we're interested in}

\item{cohort}{cohort to break out boxplots by}
}
\description{
Prints them out, used in the Rmd files
}

