% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-composites.R
\name{get_composites}
\alias{get_composites}
\title{Get composite ACE scores across module}
\usage{
get_composites(x)
}
\arguments{
\item{x}{a \code{\link{data.frame}} containing ACE summary statistics, created
using \code{\link{proc_ace_by_module}} with \code{output = "wide"}. NOTE: If processing
data containing the Filter module, this IS compatible with the \code{\link{list}}
of length 2, where the Filter data is separate, that is output by
\code{proc_ace_by_module(output = "wide")}. NOT compatible with data from \code{output = "list"}.}
}
\value{
Returns composite percentile scores for all modules present in data
 as a dataframe.
}
\description{
Function to summarize individual data across \emph{multiple modules}.
}
\section{Composite makeups}{

Composites are calculated by the following process:
1. For each module, calculate percentile rank for each individual subject
in that module's dataset
2. For each pre-specified composite group of module performance metrics,
calculate average percentile rank across each metric in the composite group.
\emph{Note:} Percentile ranks within module are calculated for all subjects
with valid data in that module, but composite rank scores between module are
calculated only for subjects who have valid data for every module required 
to calculate that composite score.

Working memory \code{wm} component metrics:
Forward and backward spatial span count, filtering capacity (k)
for 0-distractor trials only
Filtering \code{filtering} component metrics:
Filtering k for 0-distractor trials, Flanker cost of incongruence on RT,
Stroop cost of incongruence on RT
Focus \code{focus} component metrics:
Boxed cost of conjunction on RT, SAAT sustained condition RT
Impulsivity \code{impulsivity} component metrics:
SAAT impulsive condition RT
Goal management \code{goal_mgmt} component metrics:
Tap-and-trace cost of dual-task on RT, task-switch cost of switching on RT
}

