% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-composites.R
\name{get_percentile}
\alias{get_percentile}
\title{Get percentile for ONE ACE metric}
\usage{
get_percentile(df, module, id_var = "pid", var, reverse = FALSE,
  norm_dist = "self", norm_dist_grade = NULL)
}
\arguments{
\item{module}{user-specified string to be appended in the output colname}

\item{id_var}{name of subject identifier col in df
Note that this function does NOT allow a grouping variable; please only feed
data from ONE GROUP AT A TIME into this function. It's better for norming assumptions.}

\item{var}{name of data col in df. specified as a string}

\item{reverse}{should percentile scores be calculated in reverse order
(smaller number = better rank)? Defaults to \code{FALSE}.}

\item{norm_dist}{which data should the percentiles be normed to? one of the following
\code{"self"}: no norming distribution, percentiles based on data being fed in
\code{"ace_t1"}: based on results from ACE Fall 2016 data
\code{"ace_t3"}: based on results from ACE Fall 2017 data}

\item{norm_dist_grade}{if specifying an ACE norming distribution, which grade's norms?
Grades in T1: "third", "fifth", "seventh".
Grades in T3: "fourth", "sixth", "eighth". Specify as ONE OF THESE STRINGS!
Will throw error if grade not specified properly}
}
\value{
df containing \code{id_var}, \code{group_var} (if applicable), \code{var}, and percentile rank
}
\description{
Function to calculate percentile ranks for summarized data from ONE ACE metric.
}
\keyword{internal}
