% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-turns.R
\name{identify_turns}
\alias{identify_turns}
\title{Identify turn trials}
\usage{
identify_turns(y)
}
\arguments{
\item{y}{a list of "1" & "0"s}
}
\value{
Returns the indices of all turn trials.
}
\description{
Given an list of "1" & "0"'s', identify indices where a "1" is preceded by a "0".
}
\keyword{internal}
