% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ace-bulk.R
\name{load_ace_bulk}
\alias{load_ace_bulk}
\title{Read & Load all ACE csv & xls files in a directory}
\usage{
load_ace_bulk(
  path = ".",
  verbose = TRUE,
  recursive = TRUE,
  exclude = c(),
  pattern = "",
  which_modules = "",
  data_type = c("nexus", "explorer")
)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link[base]{getwd}()}.  Tilde
    expansion (see \code{\link[base]{path.expand}}) is performed.  Missing
    values will be ignored.  Elements with a marked encoding will
    be converted to the native encoding (and if that fails, considered
    non-existent).}

\item{verbose}{logical. Print details? Defaults to \code{TRUE}}

\item{recursive}{logical. Load files in subfolders also? Defaults to \code{TRUE}}

\item{exclude}{a list of patterns to exclude}

\item{pattern}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{which_modules}{Specify modules to process. Defaults to all modules.}

\item{data_type}{character What app data export type produced this data? One of
\code{c("nexus", "explorer")}. Must be specified.}
}
\value{
Returns a data.frame containing the content of every file in the
 specified \code{path}.
}
\description{
Wrapper function around \code{\link{load_ace_file}()} to read & parse 
 all ACE csv files in a directory.
}
