% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ace-demographics.R
\name{load_ace_demographics}
\alias{load_ace_demographics}
\title{Load demographics file from source}
\usage{
load_ace_demographics(file)
}
\arguments{
\item{file}{The path to the demographics file name.}
}
\value{
Returns demographic data
}
\description{
Load demographics file from source
}
