% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ace.R
\name{load_ace_file}
\alias{load_ace_file}
\title{Reads raw ACE data from a file.}
\usage{
load_ace_file(file, data_type)
}
\arguments{
\item{file}{The name of the file which the data is to be read from.}

\item{data_type}{character What app data export type produced this data? One of
\code{c("nexus", "explorer")}. Must be specified.}
}
\value{
Returns the file's content as an R \code{\link{data.frame}}.
}
\description{
Reads, parses, and converts an ACE csv or xls into an R \code{\link{data.frame}}.
}
