% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-csv.R
\name{load_sea_csv}
\alias{load_sea_csv}
\title{Because the "Reading Fluency" module has some cells with unquoted commas,
They are causing the usual read.csv delimiter guessing to split
one cell into two, creating too many columns for just a few rows}
\usage{
load_sea_csv(file)
}
\description{
Because the "Reading Fluency" module has some cells with unquoted commas,
They are causing the usual read.csv delimiter guessing to split
one cell into two, creating too many columns for just a few rows
}
\keyword{internal}
