% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sea.R
\name{load_sea_file}
\alias{load_sea_file}
\title{Reads raw SEA data from a file.}
\usage{
load_sea_file(file, verbose = FALSE)
}
\arguments{
\item{file}{The name of the file which the data is to be read from.}

\item{verbose}{Print file name (as progress checker)? Defaults to \code{FALSE}.}
}
\value{
Returns the file's content as an R \code{\link{data.frame}}.
}
\description{
Reads, parses, and converts an SEA csv into an R \code{\link{data.frame}}.
}
