% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-box.R
\name{make_box_plot}
\alias{make_box_plot}
\title{Make a box plot}
\usage{
make_box_plot(df, x, y, title, xlab, ylab, ...)
}
\arguments{
\item{df}{a data.frame}

\item{x}{the name of the x variables}

\item{y}{the name of the y variables}

\item{title}{the plot title}

\item{xlab}{the x label. defaults to x variable.}

\item{ylab}{the y label. defaults to y variable.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
Makes a box plot
}

