% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-box.R
\name{make_box_plot_bulk}
\alias{make_box_plot_bulk}
\title{Make box plots in bulk}
\usage{
make_box_plot_bulk(df, x, y = c(), prefix = "", path = NULL)
}
\arguments{
\item{x}{the name of x variable}

\item{y}{a character vector containing a list of y variables to make plots of}

\item{prefix}{Added to every plot title and file name if \code{!is.null(path)}.}
}
\description{
A wrapper function around \code{\link{make_box_plot}}.
}

