% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-write.R
\name{make_directory}
\alias{make_directory}
\title{Make a directory}
\usage{
make_directory(path, showWarnings = FALSE)
}
\arguments{
\item{path}{a character vector containing a single path name.  Tilde
    expansion (see \code{\link[base]{path.expand}}) is done.}

\item{showWarnings}{logical; should the warnings on failure be shown?}
}
\description{
Create a directory if it doesn't exist
}
