% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rlang.R
\name{map_call2_rel}
\alias{map_call2_rel}
\title{Currently only produces a set of relational statements
such as those that might be !!!-spliced into a filter call.}
\usage{
map_call2_rel(.fn, a, b, include_nas = FALSE)
}
\description{
Currently only produces a set of relational statements
such as those that might be !!!-spliced into a filter call.
}
\keyword{internal}
