% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nest.R
\name{nest_ace_raw}
\alias{nest_ace_raw}
\title{Re-nest unnested loaded ACE data}
\usage{
nest_ace_raw(df, app_type = c("classroom", "explorer"))
}
\arguments{
\item{df}{a nested \code{\link{data.frame}} containing formatted
trialwise ACE data.

This includes data loaded with the following methods: 
\enumerate{
  \item \code{\link{load_ace_file}}
  \item \code{\link{load_ace_bulk}}
}}

\item{app_type}{character. What app type produced this data? One of
\code{c("classroom", "explorer")}. Must be specified.}
}
\value{
Returns a data.frame in the nested format output by
\code{\link{load_ace_bulk}()} with the argument\code{app_type} set
to \code{"email"} or \code{"pulvinar"} for ACE Classroom data. Suitable
for further processing with, for example, \code{link{proc_by_module}()}.
}
\description{
Function to re-nest data loaded with \code{\link{load_ace_bulk}()} and
unnested with \code{link{unnest_ace_raw}()} into a form suitable
for processing in other package functions.
}
