% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-post.R
\name{post_reduce_cols}
\alias{post_reduce_cols}
\title{Subset columns from processed ACE/SEA modules}
\usage{
post_reduce_cols(
  df,
  demo_names = c("pid", "age", "grade", "gender"),
  metric_names = c(),
  metric_names_exclude = c()
)
}
\arguments{
\item{df}{a df, output by \code{\link{proc_by_module}}, containing processed
ACE or SEA data.}

\item{demo_names}{a character vector containing the names of demographic columns
to include in the output. Column names will be matched exactly.}

\item{metric_names}{a character vector containing partial names of metric columns
to \emph{include} in the output. All column names containing any of the inputs
will be included in output.}

\item{metric_names_exclude}{a character vector containing partial names of metric
columns to \emph{exclude} from the output. All column names containing any of the
inputs will be excluded from output. This argument supersedes the prior, so any
columns matching both a partial name from \code{metric_names} and a partial name
from \code{metric_names_exclude} will be excluded.}
}
\value{
a df, similar in structure to \code{proc}, but containing only the
subsetted columns.
}
\description{
User-friendly wrapper to choose demographics and selected
outcome variables from ACE/SEA data processed with \code{\link{proc_by_module}}.
}
