% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{proc_ace_complete}
\alias{proc_ace_complete}
\title{Read, Load, and Process all ACE csv files in a directory}
\usage{
proc_ace_complete(
  path_in,
  path_out = paste(path_in, "..", sep = "/"),
  verbose = TRUE,
  data_type = c("explorer", "email", "pulvinar"),
  post_metric_names = c("BRT.rt_mean.correct", "SAAT.rt_mean.correct",
    "TNT.rt_mean.correct", "object_count_span.overall", "FILTER.k", "rcs.overall")
)
}
\arguments{
\item{path_in}{character. A valid path to a folder containing exclusively raw
ACE data.}

\item{path_out}{character. A valid path to a folder containing exclusively raw
ACE data. Defaults to to \code{path_in/..}, such that the output data will be written
to the \emph{enclosing} folder containing the \code{path_in} folder. By default, this
keeps the summary data separate from the subfolder containing the raw data,
but you can specify the output path to whatever you choose
(including identical to the input path). Set this to \code{NULL} to process data
\emph{without} writing out a CSV.}

\item{verbose}{logical. Print details? Defaults to \code{TRUE}.}

\item{data_type}{character What app data export type produced this data? One of
\code{c("explorer", "email", "pulvinar")}. Must be specified.}

\item{post_metric_names}{a character vector containing \emph{partial} names of
metric columns to include in the processed output. All column names containing
any of the inputs will be included in output.}
}
\value{
Returns a \code{\link[tibble]{tibble}} containing a cleaned selection of
summary statistics from every module in the data.
}
\description{
Overarching wrapper function to read & parse 
 all ACE csv files in a directory.
}
