% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-proc.R
\name{proc_by_module}
\alias{proc_by_module}
\title{Process ACE/SEA data by module}
\usage{
proc_by_module(
  df,
  app_type = c("nexus", "explorer", "sea"),
  modules = "all",
  output = "wide",
  verbose = TRUE
)
}
\arguments{
\item{df}{a \code{\link{data.frame}} containing formatted trialwise ACE data. 

This includes data loaded with the following methods: 
\enumerate{
  \item \code{\link{load_ace_file}}
  \item \code{\link{load_ace_bulk}}
}}

\item{app_type}{character. What app type produced this data? One of
\code{c("classroom", "explorer", "sea")}. Must be specified.}

\item{modules}{character vector. Specify the names of modules (proper naming convention!)
to output data for. Defaults to all modules detected in data.}

\item{output}{string indicating preferred output format. Can be \code{"wide"} (default),
where one dataframe is output containing cols with data from all modules, or \code{"long"},
 where a dataframe is output, with a list-column containing dataframes with each module's data.}

\item{verbose}{logical. Print details? Defaults to \code{FALSE}.}
}
\value{
Returns summary statistics for every unique module included in the 
data as a list. Throws warnings for modules with undefined methods. 
See \code{\link{ace_procs}} for a list of supported modules.
}
\description{
Applies corresponding \code{\link{ace_procs}} to every unique module.
}
\section{Assumptions}{

Assumes the column \emph{module} exists in the \code{\link{data.frame}}.
}

