% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-post.R
\name{proc_na_all_by_some_cols}
\alias{proc_na_all_by_some_cols}
\title{Helper function that takes long or long-ish ACE proc data
and renders all non-demographic columns as NA for those rows
that satisfy a condition
expects filter_cols, call_filter_cols_bad, call_filter_cols_good, non_demo_cols}
\usage{
proc_na_all_by_some_cols(df)
}
\description{
Helper function that takes long or long-ish ACE proc data
and renders all non-demographic columns as NA for those rows
that satisfy a condition
expects filter_cols, call_filter_cols_bad, call_filter_cols_good, non_demo_cols
}
\keyword{internal}
