% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-post.R
\name{proc_wide_to_long}
\alias{proc_wide_to_long}
\title{Most module-post functions work much better with long proc data.
This will be called under the hood to coerce wide data to long-ish
Because it seems that users prefer the wide data}
\usage{
proc_wide_to_long(df)
}
\description{
Most module-post functions work much better with long proc data.
This will be called under the hood to coerce wide data to long-ish
Because it seems that users prefer the wide data
}
\keyword{internal}
