% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ace.R
\name{read_ace_file}
\alias{read_ace_file}
\title{Reads raw ace data from a file.}
\usage{
read_ace_file(file)
}
\arguments{
\item{file}{The name of the file which the data is to be read from.}
}
\value{
Returns the file's content as an R \code{\link{data.frame}}.
}
\description{
Reads, parses, and converts an ACE csv or xls into an R \code{\link{data.frame}}.
}

