% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load-csv.R
\name{read_raw_csv}
\alias{read_raw_csv}
\title{Read raw csv data}
\usage{
read_raw_csv(file)
}
\arguments{
\item{file}{The name of the file which the data is to be read from.}
}
\value{
Returns the file's content as an R \code{\link{data.frame}}.
}
\description{
Reads, parses, and converts an ACE csv into an R \code{\link{data.frame}}.
}

