% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load-data-csv.R
\name{read_raw_csv_in_directory}
\alias{read_raw_csv_in_directory}
\title{Read & Load all ACE csv files in a directory}
\usage{
read_raw_csv_in_directory(path = ".", pattern = NULL, verbose = TRUE)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link{getwd}()}.  Tilde
    expansion (see \code{\link{path.expand}}) is performed.  Missing
    values will be ignored.}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{verbose.}{logical. Print details? Defaults to \code{TRUE}}
}
\value{
Returns a data.frame containing the content of every file in the
 specified \code{path}.
}
\description{
Wrapper function around \code{\link{read_raw_csv}()}. to read & parse
 all ACE csv files in a directory.
}

