% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils-clean.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Remove empty columns from a data frame.}
\usage{
remove_empty_cols(dat)
}
\arguments{
\item{dat}{a data frame}
}
\value{
Returns a data frame where the empty columns
 have been removed
}
\description{
Remove empty columns from a data frame.
}

