% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils-clean.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Remove empty rows from a data frame.}
\usage{
remove_empty_rows(dat)
}
\arguments{
\item{dat}{a data frame}
}
\value{
Returns a data frame where the empty rows have been removed
}
\description{
Remove empty rows from a data frame.
}

