% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils-clean.R
\name{remove_rows}
\alias{remove_rows}
\title{Remove specified rows from a data frame}
\usage{
remove_rows(dat, rows = c(), reset = TRUE)
}
\arguments{
\item{dat}{a data frame}

\item{rows}{a vector containing the names of the the rows to remove}

\item{reset}{reset the row numbers?}
}
\value{
Returns a data frame where the specified rows have been removed
}
\description{
Remove specified rows from a data frame
}

