% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rename.R
\name{rename_overwrite}
\alias{rename_overwrite}
\title{Renames tibble column "x" to "y", dropping the old contents of "y"}
\usage{
rename_overwrite(df, col_rename, col_drop)
}
\description{
Renames tibble column "x" to "y", dropping the old contents of "y"
}
\keyword{internal}
