% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sea-constant-columns.R
\docType{data}
\name{sea_header}
\alias{sea_header}
\alias{COL_RESPONSE}
\alias{COL_QUESTION_TEXT}
\alias{COL_MODULE}
\alias{COL_SEA_GROUP}
\alias{ALL_POSSIBLE_SEA_DEMOS}
\alias{Q_ALL_POSSIBLE_SEA_DEMOS}
\alias{standardize_sea_column_names}
\alias{standardize_sea_values}
\title{SEA column name constants}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 9.

An object of class \code{list} of length 9.
}
\usage{
COL_RESPONSE

COL_QUESTION_TEXT

COL_MODULE

COL_SEA_GROUP

ALL_POSSIBLE_SEA_DEMOS

Q_ALL_POSSIBLE_SEA_DEMOS

standardize_sea_column_names(df)

standardize_sea_values(df)
}
\description{
SEA also uses some ACE column name constants.
These are constants which are SEA-specific.
}
\keyword{datasets}
\keyword{internal}
