% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sea-constants-modules.R
\docType{data}
\name{sea_module}
\alias{sea_module}
\alias{sea_module}
\alias{ALL_SEA_MODULES}
\title{SEA module constants}
\format{An object of class \code{NULL} of length 0.}
\usage{
ALL_SEA_MODULES
}
\description{
SEA module constants
}
\keyword{datasets}
\keyword{internal}
