% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-seacrest-standardize.R
\name{standardize_seacrest_pid}
\alias{standardize_seacrest_pid}
\title{Standardize Seacrest PIDs}
\usage{
standardize_seacrest_pid(x)
}
\arguments{
\item{x}{A character vector containing raw seacrest pids}
}
\description{
Standardizes Seacrest PIDs.
}

