% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-trim.R
\name{trim_rt_trials}
\alias{trim_rt_trials}
\title{Trim trials from ACE/SEA data by reaction time}
\usage{
trim_rt_trials(df, sd_cutoff = FALSE, range_cutoff = FALSE,
  verbose = TRUE)
}
\arguments{
\item{df}{a \code{\link{data.frame}} containing formatted trialwise ACE data. 

This includes data loaded with the following methods: 
\enumerate{
  \item \code{\link{load_ace_file}}
  \item \code{\link{load_ace_bulk}}
}}

\item{sd_cutoff}{numeric. Remove within-subject RTs further than this many SD from
within-subject mean RT? Enter as one number. If both range and SD are specified,
will apply range cutoff \emph{first,} and then will apply SD cutoff \emph{afterward}.
Defaults to \code{FALSE}.}

\item{range_cutoff}{numeric vector, length 2. Remove within-subject RTs outside of
this specified range? Enter min and max accepted RTs as a vector length 2. If min or max
not specified, enter that value as NA in the vector. Defaults to \code{FALSE}.}

\item{verbose}{logical. Print details? Defaults to \code{TRUE}.}
}
\value{
Returns the input data, with RTs corresponding to offending trials rendered as NA.
}
\description{
Applies corresponding \code{\link{ace_trims}} to every session of data.
}
\section{Assumptions}{

Assumes the \code{\link{data.frame}} is nested, with two columns:
\code{module} (character) and \code{data} (list, each containing a \code{\link{data.frame}}).
}

