% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-trim.R
\name{trim_rt_trials_range}
\alias{trim_rt_trials_range}
\title{Trim trials from ACE/SEA data by response time within a fixed range}
\usage{
trim_rt_trials_range(
  df,
  cutoff_min = NA,
  cutoff_max = NA,
  exclude = c(),
  verbose = TRUE
)
}
\arguments{
\item{df}{a \code{\link{data.frame}} containing formatted trialwise ACE data. 

This includes data loaded with the following methods: 
\enumerate{
  \item \code{\link{load_ace_file}}
  \item \code{\link{load_ace_bulk}}
}}

\item{cutoff_min}{numeric. Remove within-subject RTs \emph{below} (but not equal to)
this specified value (in ms)? Defaults to \code{NA}, where no too-low values will be scrubbed.
Note that \code{\link{load_ace_bulk}} already removes RTs below 150 ms for all modules,
so those can never be included even if this minimum cutoff is set to a value below 150 ms.}

\item{cutoff_max}{numeric vector. Remove within-subject RTs \emph{above} (but not equal to)
this specified value? Defaults to \code{NA}, where no too-high values will be scrubbed.
Defaults to \code{FALSE}.}

\item{exclude}{character vector. Specify the names of modules (proper naming convention!)
that should be \emph{ignored}. Defaults to \code{c()}, where all modules are subject to
range scrubbing. Note that this function always excludes spatial span tasks, as those are 
much shorter than other modules.}

\item{verbose}{logical. Print details? Defaults to \code{TRUE}.}
}
\value{
Returns the input data, with RTs corresponding to offending trials rendered as NA.
}
\description{
Applies corresponding \code{\link{ace_trims}} to every session of data.
}
\section{Assumptions}{

Assumes the \code{\link{data.frame}} is nested, with two columns:
\code{module} (character) and \code{data} (list, each containing a \code{\link{data.frame}}).
}

