% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-trim.R
\name{trim_rt_trials_sd}
\alias{trim_rt_trials_sd}
\title{Trim trials from ACE/SEA data by response time within N SDs of the mean}
\usage{
trim_rt_trials_sd(
  df,
  cutoff = 3,
  exclude = c("SAAT", "SAATIMPULSIVE", "SAATSUSTAINED"),
  verbose = TRUE
)
}
\arguments{
\item{df}{a \code{\link{data.frame}} containing formatted trialwise ACE data. 

This includes data loaded with the following methods: 
\enumerate{
  \item \code{\link{load_ace_file}}
  \item \code{\link{load_ace_bulk}}
}}

\item{cutoff}{numeric. Remove within-subject RTs further than this many SD from
within-subject mean RT? Enter as one number. Defaults to 3.}

\item{exclude}{character vector. Specify the names of modules (proper naming convention!)
that should be \emph{ignored}. Defaults to all \code{"SAAT"} data, but can be specified as an empty
character vector (in which case the SAAT module's RTs would be subject to SD scrubbing).
Note that this function always excludes spatial span tasks, as those are much shorter 
than other modules.}

\item{verbose}{logical. Print details? Defaults to \code{TRUE}.}
}
\value{
Returns the input data, with RTs corresponding to offending trials rendered as NA.
}
\description{
Applies corresponding \code{\link{ace_trims}} to every session of data.
If trimming is to be done by range and by SD, we recommend calling \code{\link{trim_rt_trials_range}}
first, and then calling this function to trim by SD.
}
\section{Assumptions}{

Assumes the \code{\link{data.frame}} is nested, with two columns:
\code{module} (character) and \code{data} (list, each containing a \code{\link{data.frame}}).
}

