% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nest.R
\name{unnest_ace_raw}
\alias{unnest_ace_raw}
\title{Unnest loaded ACE data into long form}
\usage{
unnest_ace_raw(df, app_type = c("classroom", "explorer"))
}
\arguments{
\item{df}{a nested \code{\link{data.frame}} containing formatted
trialwise ACE data. 

This includes data loaded with the following methods: 
\enumerate{
  \item \code{\link{load_ace_file}}
  \item \code{\link{load_ace_bulk}}
}}

\item{app_type}{character. What app type produced this data? One of
\code{c("classroom", "explorer")}. Must be specified.}
}
\value{
Returns a data.frame with no list-columns, with one row per trial,
and one section of rows per module. Columns only containing valid data
for one or a subset of modules are all set to \code{NA} for modules to which
those columns do not apply.
}
\description{
Function to unnest data loaded with \code{\link{load_ace_bulk}()}
into long form.
}
