% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-constant-columns.R
\docType{data}
\name{ace_header}
\alias{COL_CONDITION}
\alias{COL_CORRECT_BUTTON}
\alias{COL_CORRECT_RESPONSE}
\alias{COL_PID}
\alias{COL_RT}
\alias{COL_RW}
\alias{COL_TRIAL_TYPE}
\alias{ace_header}
\alias{standardize_ace_column_names}
\title{ACE column name constants}
\format{An object of class \code{character} of length 1.}
\usage{
COL_RT

COL_CORRECT_BUTTON

COL_CORRECT_RESPONSE

COL_RW

COL_CONDITION

COL_PID

COL_TRIAL_TYPE

standardize_ace_column_names(df)
}
\description{
ACE column name constants
}
\keyword{datasets}
\keyword{internal}

