% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-constant-columns.R
\docType{data}
\name{ace_header}
\alias{COL_ACC}
\alias{COL_RT}
\alias{COL_RW}
\alias{ace_header}
\alias{standardize_ace_column_names}
\title{ACE column name constants}
\format{An object of class \code{character} of length 1.}
\usage{
COL_RT

COL_ACC

COL_RW

standardize_ace_column_names(df)
}
\description{
ACE column name constants
}
\keyword{datasets}
\keyword{internal}

