% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load-constants.R
\docType{data}
\name{ace_module}
\alias{BOXED}
\alias{BRT}
\alias{DISCRIMINATION}
\alias{FLANKER}
\alias{SAAT}
\alias{SPATIAL_SPAN}
\alias{STROOP}
\alias{TASK_SWITCH}
\alias{TNT}
\alias{ace_module}
\title{ACE module constants}
\format{\preformatted{ chr "BOXED"
}}
\usage{
BOXED

BRT

DISCRIMINATION

FLANKER

SAAT

SPATIAL_SPAN

STROOP

TASK_SWITCH

TNT
}
\description{
ACE module constants
}
\keyword{datasets}
\keyword{internal}

