% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-constants.R
\docType{data}
\name{ace_module_constants}
\alias{PARTICIPANT_BY_GROUP}
\alias{PBG}
\alias{ace_module_constants}
\title{ACE module constants}
\format{An object of class \code{character} of length 2.}
\usage{
PARTICIPANT_BY_GROUP

PBG
}
\description{
ACE module constants
}
\keyword{datasets}
\keyword{internal}

