% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-module-boxed.R, R/module-proc.R
\name{ace_procs}
\alias{ace_procs}
\alias{module_boxed}
\title{ACE module procs}
\usage{
module_boxed(df)
}
\description{
Methods for processing user data by \emph{group}.
}
\keyword{internal}

