% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-modules.R, R/module-proc.R
\name{ace_procs}
\alias{ace_procs}
\alias{module_boxed}
\alias{module_brt}
\alias{module_discrimination}
\alias{module_flanker}
\alias{module_saat}
\alias{module_spatialspan}
\alias{module_stroop}
\alias{module_taskswitch}
\alias{module_tnt}
\title{ACE module procs}
\usage{
module_boxed(df)

module_brt(df)

module_discrimination(df)

module_flanker(df)

module_saat(df)

module_stroop(df)

module_spatialspan(df)

module_taskswitch(df)

module_tnt(df)
}
\description{
Methods for processing user data by \emph{group}.
}
\keyword{internal}

