% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aceR-package.R
\name{aceR_sample_data_path}
\alias{aceR_sample_data_path}
\title{aceR sample data path}
\usage{
aceR_sample_data_path()
}
\value{
Returns the path containing sample aceR data
}
\description{
aceR sample data path
}

