% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-composites.R
\name{get_all_percentiles}
\alias{get_all_percentiles}
\title{Get ALL percentiles for available ACE metrics}
\usage{
get_all_percentiles(data, norm_dist = "self", norm_dist_grade = NULL)
}
\arguments{
\item{df}{Processed ACE data. Expects "wide" output ACE data from proc_by_module()}
}
\value{
a df with PID, valid demographics, and percentile rank for the specified norming distribution
see \code{get_percentile()} for info on norming distributions
}
\description{
Function to return percentile ranks for ACE metrics from \emph{multiple modules}.
}
