% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-constants.R
\name{identify_module}
\alias{identify_module}
\title{Identify ACE module from filename}
\usage{
identify_module(file)
}
\arguments{
\item{file}{The name of the file which the data is to be read from.}
}
\value{
Returns the name of the ACE module if found.
}
\description{
Identifies ACE module from the filename
}

