% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-generic.R
\name{load_files}
\alias{load_files}
\title{Load all files in a directory as a data frame}
\usage{
load_files(path = ".", verbose = FALSE, ...)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link[base]{getwd}()}.  Tilde
    expansion (see \code{\link[base]{path.expand}}) is performed.  Missing
    values will be ignored.  Elements with a marked encoding will
    be converted to the native encoding (and if that fails, considered
    non-existent).}

\item{verbose}{print details?}

\item{...}{see \code{\link{list.files}}}
}
\value{
all files in a directory as one data frame
}
\description{
Load all files in a directory as one data frame
}
\section{Warning}{

Use \code{\link{load_ace_bulk}} for raw formatted ACE data.
}

