% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-generic.R
\name{load_files}
\alias{load_files}
\title{Load all files in a directory as a data frame}
\usage{
load_files(path = ".", verbose = FALSE, ...)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link{getwd}()}.  Tilde
    expansion (see \code{\link{path.expand}}) is performed.  Missing
    values will be ignored.}
}
\value{
all files in a directory as one data frame
}
\description{
Load all files in a directory as one data frame
}
\section{Warning}{

Use \code{\link{load_ace_bulk}} for raw formatted ACE data.
}

