% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils-filter-df.R
\name{numeric_row_names}
\alias{numeric_row_names}
\title{Numeric row names}
\usage{
numeric_row_names(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
Returns numeric row names
}
\description{
Numeric row names
}

