% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-post.R
\name{post_clean_chance}
\alias{post_clean_chance}
\title{Scrub processed data with below-chance accuracy}
\usage{
post_clean_chance(
  df,
  overall = TRUE,
  cutoff_dprime = 0,
  cutoff_2choice = 0.5,
  cutoff_4choice = 0.25
)
}
\arguments{
\item{df}{a df, output by \code{\link{proc_by_module}}, containing processed
ACE or SEA data.}

\item{overall}{Also scrub ".overall" data? Defaults to \code{TRUE}.}

\item{cutoff_dprime}{Minimum value of d' to allow, for relevant tasks
(Tap and Trace, SAAT, Filter). Defaults to 0.}

\item{cutoff_2choice}{Minimum value of accuracy to allow, for 2-response tasks
(Flanker, Boxed). Defaults to 0.5.}

\item{cutoff_4choice}{Minimum value of accuracy to allow, for 4-response tasks
(Stroop, Task Switch). Defaults to 0.25.}
}
\value{
a df, similar in structure to \code{proc}, but with below-cutoff values in
certain columns converted to \code{NA}.
}
\description{
User-friendly wrapper to replace below-chance records with \code{NA}
in ACE/SEA data processed with \code{\link{proc_by_module}}.
}
