% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-post.R
\name{post_clean_low_trials}
\alias{post_clean_low_trials}
\title{Scrub processed data with too few trials}
\usage{
post_clean_low_trials(df, min_trials = 5)
}
\arguments{
\item{df}{a df, output by \code{\link{proc_by_module}}, containing processed
ACE or SEA data.}

\item{min_trials}{Minimum number of trials to require in most restrictive condition.
Defaults to 5.}
}
\value{
a df, similar in structure to \code{proc}, but with records with too few trials
converted to \code{NA}.
}
\description{
User-friendly wrapper to replace below-chance records with \code{NA}
in ACE/SEA data processed with \code{\link{proc_by_module}}.
}
