% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{proc_ace_complete}
\alias{proc_ace_complete}
\title{Read, Load, and Process all ACE csv files in a directory}
\usage{
proc_ace_complete(
  path_in,
  path_out = paste(path_in, "..", sep = "/"),
  verbose = TRUE,
  data_type = c("explorer", "email", "pulvinar"),
  rt_cutoff_min = 200,
  post_min_trials = 5,
  post_chance_level = "overall",
  post_metric_names = c("BRT.rt_mean.correct", "SAATIMPULSIVE.rt_mean.correct",
    "SAATSUSTAINED.rt_mean.correct", "TNT.rt_mean.correct", "object_count_span.overall",
    "FILTER.k", "COLORSELECTION.max_delay_time.correct.strict", "rcs.overall"),
  post_metric_names_exclude = c("FILTER.rcs", "TNT.rt_mean.correct.")
)
}
\arguments{
\item{path_in}{character. A valid path to a folder containing exclusively raw
ACE data.}

\item{path_out}{character. A valid path to a folder containing exclusively raw
ACE data. Defaults to to \code{path_in/..}, such that the output data will be written
to the \emph{enclosing} folder containing the \code{path_in} folder. By default, this
keeps the summary data separate from the subfolder containing the raw data,
but you can specify the output path to whatever you choose
(including identical to the input path). Set this to \code{NULL} to process data
\emph{without} writing out a CSV.}

\item{verbose}{logical. Print details? Defaults to \code{TRUE}.}

\item{data_type}{character What app data export type produced this data? One of
\code{c("explorer", "email", "pulvinar")}. Must be specified.}

\item{rt_cutoff_min}{numeric. Remove within-subject RTs \emph{below} (but not equal to)
this specified value (in ms)? Defaults to \code{200}. Can be set to \code{NA},
where no too-low values will be scrubbed.
Note that \code{\link{load_ace_bulk}} already removes RTs below 150 ms for all modules,
so those can never be included even if this minimum cutoff is set to a value below 150 ms.
Passed through to \code{\link{trim_rt_trials_range}}.}

\item{post_min_trials}{Minimum number of trials to require in most restrictive condition.
Defaults to 5. This condition is checked against the \code{*_count} summary columns,
that count all trials with a valid response time (and all no-go trials, if a response
was not expected.) Passed through to \code{\link{post_clean_low_trials}}. Set this argument
to \code{NA} to leave all trials in and skip this step.}

\item{post_chance_level}{character. How strictly to retain records containing
above-threshold performance? One of:
\itemize{
  \item \code{"overall"}: Trim based on \emph{whole-module} accuracy. Minimum of: d' > 0,
  2-choice task accuracy > 0.5, 4-choice task accuracy > 0.25. \strong{Default.}
  \item \code{"easy"}: Trim based on \emph{within-condition} accuracy. Minimum of: d' > 0,
  2-choice task accuracy > 0.5, 4-choice task accuracy > 0.25.
  \item \code{"none"}: Do not trim any records by performance threshold.
}
Passed through to \code{\link{post_clean_chance}}.}

\item{post_metric_names}{a character vector containing \emph{partial} names of
metric columns to include in the processed output. All column names containing
any of the inputs will be included in output. Passed through to \code{\link{post_reduce_cols}}.}

\item{post_metric_names_exclude}{a character vector containing \emph{partial} names of
metric columns to exclude from  the processed output. All column names containing
any of the inputs will be excluded from output. Passed through to \code{\link{post_reduce_cols}}.}
}
\value{
Returns a \code{\link[tibble]{tibble}} containing a cleaned selection of
summary statistics from every module in the data.
}
\description{
Overarching wrapper function to read & parse 
 all ACE csv files in a directory.
}
