% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-proc.R
\name{proc_by_module}
\alias{proc_by_module}
\title{Process ACE data by module}
\usage{
proc_by_module(df)
}
\arguments{
\item{df}{a \code{\link{data.frame}} containing formatted ACE data. 

This includes data loaded with the following methods: 
\enumerate{
  \item \code{\link{read_raw_csv_in_directory}}
}}
}
\value{
Returns summary statistics for every unique module included in the 
 data as a list. Throws warnings for modules with undefined methods. 
 See \code{\link{ace_procs}} for a list of supported modules.
}
\description{
Applies corresponding \code{\link{ace_procs}} to every unique module.
}
\section{Assumptions}{

Assumes the column \emph{module} exists in the \code{\link{data.frame}}.
}

