% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-proc.R
\name{proc_by_module}
\alias{proc_by_module}
\title{Process ACE data by module}
\usage{
proc_by_module(df, scrub_short = TRUE, conditions = NULL, verbose = FALSE)
}
\arguments{
\item{df}{a \code{\link{data.frame}} containing formatted ACE data. 

This includes data loaded with the following methods: 
\enumerate{
  \item \code{\link{load_ace_file}}
  \item \code{\link{load_ace_bulk}}
}}

\item{scrub_short}{logical. Remove subjects with <1/2 of trials? Defaults to \code{TRUE}}

\item{conditions}{character vector. If data contains multiple study conditions
(e.g. pre & post), specify their labels here. Case insensitive.}

\item{verbose}{logical. Print details? Defaults to \code{FALSE}.}
}
\value{
Returns summary statistics for every unique module included in the 
 data as a list. Throws warnings for modules with undefined methods. 
 See \code{\link{ace_procs}} for a list of supported modules.
}
\description{
Applies corresponding \code{\link{ace_procs}} to every unique module.
}
\section{Assumptions}{

Assumes the column \emph{module} exists in the \code{\link{data.frame}}.
}

