% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-proc.R
\name{proc_by_module}
\alias{proc_by_module}
\title{Process ACE/SEA data by module}
\usage{
proc_by_module(df, modules = "all", output = "wide",
  conditions = NULL, verbose = TRUE)
}
\arguments{
\item{df}{a \code{\link{data.frame}} containing formatted trialwise ACE data. 

This includes data loaded with the following methods: 
\enumerate{
  \item \code{\link{load_ace_file}}
  \item \code{\link{load_ace_bulk}}
}}

\item{modules}{character vector. Specify the names of modules (proper naming convention!)
to output data for. Defaults to all modules detected in data.}

\item{output}{string indicating preferred output format. Can be \code{"wide"} (default),
where one dataframe is output containing cols with data from all modules, or \code{"long"},
 where a dataframe is output, with a list-column containing dataframes with each module's data.}

\item{conditions}{character vector. If data contains multiple study conditions
(e.g. pre & post), specify their labels here. Case insensitive.}

\item{verbose}{logical. Print details? Defaults to \code{TRUE}.}
}
\value{
Returns summary statistics for every unique module included in the 
 data as a list. Throws warnings for modules with undefined methods. 
 See \code{\link{ace_procs}} for a list of supported modules.
}
\description{
Applies corresponding \code{\link{ace_procs}} to every unique module.
}
\section{Assumptions}{

Assumes the \code{\link{data.frame}} is nested, with two columns:
\code{module} (character) and \code{data} (list, each containing a \code{\link{data.frame}}).
}

