% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-proc.R
\name{proc_by_module}
\alias{proc_by_module}
\title{Process ACE data by module}
\usage{
proc_by_module(df, modules = "all", output = "wide",
  rm_outlier_rts_sd = FALSE, rm_outlier_rts_range = FALSE,
  rm_short_subs = FALSE, conditions = NULL, verbose = FALSE)
}
\arguments{
\item{df}{a \code{\link{data.frame}} containing formatted trialwise ACE data. 

This includes data loaded with the following methods: 
\enumerate{
  \item \code{\link{load_ace_file}}
  \item \code{\link{load_ace_bulk}}
}}

\item{modules}{character vector. Specify the names of modules (proper naming convention!)
to output data for. Defaults to all modules detected in data.}

\item{output}{string indicating preferred output format. Can be \code{"wide"} (default),
where one dataframe is output containing cols with data from all modules, or \code{"long"},
 where a dataframe is output, with a list-column containing dataframes with each module's data.}

\item{rm_outlier_rts_sd}{numeric. Remove within-subject RTs further than this many SD from
within-subject mean RT? Enter as one number. Specify either this or \code{rm_outlier_rts_range},
but not both. If both specified, will use SD cutoff. Defaults to \code{FALSE}.}

\item{rm_outlier_rts_range}{numeric vector, length 2. Remove within-subject RTs outside of
this specified range? Enter min and max accepted RTs as a vector length 2. If min or max
not specified, enter that value as NA in the vector. Specify either this or \code{rm_outlier_rts_range},
but not both. If both specified, will use SD cutoff. Defaults to \code{FALSE}.}

\item{rm_short_subs}{logical. Remove subjects with <1/2 of trials? Defaults to \code{TRUE}}

\item{conditions}{character vector. If data contains multiple study conditions
(e.g. pre & post), specify their labels here. Case insensitive.}

\item{verbose}{logical. Print details? Defaults to \code{FALSE}.}
}
\value{
Returns summary statistics for every unique module included in the 
 data as a list. Throws warnings for modules with undefined methods. 
 See \code{\link{ace_procs}} for a list of supported modules.
}
\description{
Applies corresponding \code{\link{ace_procs}} to every unique module.
}
\section{Assumptions}{

Assumes the column \emph{module} exists in the \code{\link{data.frame}}.
}

