% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sea-constant-columns.R
\docType{data}
\name{sea_header}
\alias{sea_header}
\alias{sea_header}
\alias{COL_RESPONSE}
\alias{sea_header}
\alias{COL_QUESTION_TEXT}
\alias{sea_header}
\alias{COL_MODULE}
\alias{sea_header}
\alias{standardize_sea_column_names}
\title{SEA column name constants}
\format{An object of class \code{character} of length 1.}
\usage{
COL_RESPONSE

COL_QUESTION_TEXT

COL_MODULE

standardize_sea_column_names(df)
}
\description{
SEA also uses some ACE column name constants.
These are constants which are SEA-specific.
}
\keyword{datasets}
\keyword{internal}
