% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sea-constants-modules.R
\docType{data}
\name{sea_module}
\alias{sea_module}
\alias{MATH_FLU}
\alias{READ_FLU}
\alias{FRAC_1}
\alias{FRAC_2}
\alias{FRAC_3}
\alias{MATH_REC}
\alias{RUN_MEM_SPAN}
\alias{GROUPITIZE}
\alias{REL_MATCH}
\alias{ARITHM_VER}
\alias{READ_COMP}
\alias{ALL_SEA_MODULES}
\title{SEA module constants}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 11.
}
\usage{
MATH_FLU

READ_FLU

FRAC_1

FRAC_2

FRAC_3

MATH_REC

RUN_MEM_SPAN

GROUPITIZE

REL_MATCH

ARITHM_VER

READ_COMP

ALL_SEA_MODULES
}
\description{
SEA module constants
}
\keyword{datasets}
\keyword{internal}
